+++ src/agents/router.py
@@ -1,7 +1,7 @@
 import random
 from openai.types.beta.threads import Message
 from openai.pagination import SyncCursorPage
-from tools.openai_clients import litellm_client
+from utils.openai_clients import litellm_client
 
 
 class RouterAgent:
